requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.40
aiosqlite==0.20.0
alembic==1.15.2
httpx==0.28.1
beautifulsoup4==4.13.4
itsdangerous==2.2.0
python-dotenv==1.1.0
resend==2.10.0
